Add `base_revision` and `detect_rollback`
Allows providing the `base_revision` to use for scanning commits for issue keys.
This opens up support for reporting sets of changes that may be included in many
commits.

To replicate the previous behavior, one could use `<< pipeline.git.base_revision >>`
to get the commit used in the previous CircleCI job. This would ensure only new changes
from the latest sets of commits would be used.

This also opens up the behavior to support using the pipeline base_revision of the last
commit for detecting potential rollbacks. You can set `detect_rollback: true` to enable
this feature which will be detected when rerunning a past job that is using a commit
earlier in history than the latest HEAD. It will then report any issues found between
that past commit..origin/HEAD as `rolled_back` in Jira.